Version 4.6. Much better 256 terminal colors
Almost everything looks the same in a 256 color terminal and MacVim.
Title, Constant, PreProc and Special looks better in MacVim but they are
as close as they can possibly get right now.

Added a bunch of new colors and fixed the ones that only looked bad in
a 256 color terminal but not MacVim.

This colors scheme is starting to near it's final stage. Only small
fixes will be added from here on out (probably).